Moved DI to BusinessLogic
